refactor: Hart and Dominion moved to shangrla/formats; moved outdated notebooks to examples/old
