+++ shangrla/formats/__init__.py
@@ -1,0 +1,7 @@
+"""
+Functionality for reading a variety of CVR formats.
+"""
+
+__all__ = ["Dominion", "Hart"]
+
+from . import *
